Update on "[jit] split TestDataParallel, TestAsync"
Differential Revision: [D18339007](https://our.internmc.facebook.com/intern/diff/D18339007)
